refactor(#1146): Remove JavaSourceCompilationIT and package-info.java files
