--- src/test/java/it/package-info.java
@@ -1,8 +1,0 @@
-/*
- * SPDX-FileCopyrightText: Copyright (c) 2016-2025 Objectionary.com
- * SPDX-License-Identifier: MIT
- */
-/**
- * Integration tests.
- */
-package it;
